# internal/orders/domain.go
package orders

import (
	"order-sync/internal/contracts"
)

type OrderItem = contracts.OrderItem
type OrderStatus = contracts.OrderStatus
type WebhookResult = contracts.WebhookResult
type ValidationIssue = contracts.ValidationIssue